feat(RestockLog) : Add get_by_product_id
Add function to get the restock logs according to the product ID.
We will use this function in the next commit to delete the logs
when user deletes the product. Becuase the product ID in RestockLog
is a ForeignKey, we cannot delete a product without deleting the
restock logs first.

Issue: #19

Signed-off-by: Yaakov Neuman <yakinew@yahoo.com>